+++ Code/Engine/Foundation/Platform/NoImpl/Screen_NoImpl.h
@@ -3,7 +3,7 @@ EZ_FOUNDATION_INTERNAL_HEADER
 
 #include <Foundation/System/Screen.h>
 
-ezResult ezScreen::EnumerateScreens(ezHybridArray<ezScreenInfo, 2>& out_Screens)
+ezResult ezScreen::EnumerateScreens(ezDynamicArray<ezScreenInfo>& out_Screens)
 {
   EZ_ASSERT_NOT_IMPLEMENTED;
   return EZ_FAILURE;
