+++ Code/UnitTests/TestFramework/Framework/TestFramework.h
@@ -4,6 +4,7 @@
 #include <TestFramework/Framework/SimpleTest.h>
 #include <TestFramework/Framework/TestBaseClass.h>
 #include <TestFramework/Framework/TestResults.h>
+#include <TestFramework/Platform/TestFrameworkEntryPoint.h>
 #include <TestFramework/TestFrameworkDLL.h>
 
 #include <Foundation/Containers/DynamicArray.h>
@@ -222,156 +223,6 @@ protected:
   ezUInt32 m_uiCurrentSubTestIndex = ezInvalidIndex;
   bool m_bTestsRunning = false;
 };
-
-#ifdef EZ_NV_OPTIMUS
-#  undef EZ_NV_OPTIMUS
-#endif
-
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
-#  include <Foundation/Basics/Platform/Win/MinWindows.h>
-#  define EZ_NV_OPTIMUS                                                                           \
-    extern "C"                                                                                    \
-    {                                                                                             \
-      _declspec(dllexport) ezMinWindows::DWORD NvOptimusEnablement = 0x00000001;                  \
-      _declspec(dllexport) ezMinWindows::DWORD AmdPowerXpressRequestHighPerformance = 0x00000001; \
-    }
-#else
-#  define EZ_NV_OPTIMUS
-#endif
-
-#if EZ_ENABLED(EZ_PLATFORM_ANDROID)
-#  include <Foundation/Basics/Platform/Android/AndroidUtils.h>
-#  include <TestFramework/Utilities/TestSetup.h>
-#  include <android/log.h>
-#  include <android/native_activity.h>
-#  include <android_native_app_glue.h>
-
-static void ezAndroidHandleCmd(struct android_app* pApp, int32_t cmd);
-int ezAndroidMain(int argc, char** argv);
-
-class ezAndroidTestApplication
-{
-public:
-  ezAndroidTestApplication(struct android_app* pApp)
-    : m_pApp(pApp)
-  {
-    pApp->userData = this;
-    pApp->onAppCmd = ezAndroidHandleCmd;
-    ezAndroidUtils::SetAndroidApp(pApp);
-  }
-
-  void HandleCmd(int32_t cmd)
-  {
-    switch (cmd)
-    {
-      case APP_CMD_INIT_WINDOW:
-        if (m_pApp->window != nullptr)
-        {
-          ezAndroidMain(0, nullptr);
-          m_bStarted = true;
-
-          int width = ANativeWindow_getWidth(m_pApp->window);
-          int height = ANativeWindow_getHeight(m_pApp->window);
-          ezLog::Info("Init Window: {}x{}", width, height);
-        }
-        break;
-      default:
-        break;
-    }
-  }
-
-  void AndroidRun()
-  {
-    bool bRun = true;
-    while (true)
-    {
-      struct android_poll_source* pSource = nullptr;
-      int iIdent = 0;
-      int iEvents = 0;
-      while ((iIdent = ALooper_pollAll(0, nullptr, &iEvents, (void**)&pSource)) >= 0)
-      {
-        if (pSource != nullptr)
-          pSource->process(m_pApp, pSource);
-      }
-
-      // APP_CMD_INIT_WINDOW has not triggered yet. Engine is not yet started.
-      if (!m_bStarted)
-        continue;
-
-      if (bRun && ezTestSetup::RunTests() != ezTestAppRun::Continue)
-      {
-        bRun = false;
-        ANativeActivity_finish(m_pApp->activity);
-      }
-      if (m_pApp->destroyRequested)
-      {
-        break;
-      }
-    }
-  }
-
-private:
-  struct android_app* m_pApp = nullptr;
-  bool m_bStarted = false;
-};
-
-static void ezAndroidHandleCmd(struct android_app* pApp, int32_t cmd)
-{
-  ezAndroidTestApplication* pAndroidApp = static_cast<ezAndroidTestApplication*>(pApp->userData);
-  pAndroidApp->HandleCmd(cmd);
-}
-
-#  define EZ_TESTFRAMEWORK_ENTRY_POINT_BEGIN(szTestName, szNiceTestName)                                                \
-    int ezAndroidMain(int argc, char** argv);                                                                           \
-    extern "C" void android_main(struct android_app* app)                                                               \
-    {                                                                                                                   \
-      ezAndroidTestApplication androidApp(app);                                                                         \
-      androidApp.AndroidRun();                                                                                          \
-      const ezInt32 iFailedTests = ezTestSetup::GetFailedTestCount();                                                   \
-      ezTestSetup::DeInitTestFramework();                                                                               \
-      __android_log_print(ANDROID_LOG_ERROR, "ezEngine", "Test framework exited with return code: '%d'", iFailedTests); \
-    }                                                                                                                   \
-                                                                                                                        \
-    int ezAndroidMain(int argc, char** argv)                                                                            \
-    {                                                                                                                   \
-      ezTestSetup::InitTestFramework(szTestName, szNiceTestName, 0, nullptr);                                           \
-      /* Execute custom init code here by using the BEGIN/END macros directly */
-
-#else
-/// \brief Macro to define the application entry point for all test applications
-#  define EZ_TESTFRAMEWORK_ENTRY_POINT_BEGIN(szTestName, szNiceTestName)                    \
-    /* Enables that on machines with multiple GPUs the NVIDIA GPU is preferred */           \
-    EZ_NV_OPTIMUS                                                                           \
-    EZ_APPLICATION_ENTRY_POINT_CODE_INJECTION                                               \
-    int main(int argc, char** argv)                                                         \
-    {                                                                                       \
-      ezTestSetup::InitTestFramework(szTestName, szNiceTestName, argc, (const char**)argv); \
-      /* Execute custom init code here by using the BEGIN/END macros directly */
-
-#endif
-
-#if EZ_ENABLED(EZ_PLATFORM_ANDROID)
-#  define EZ_TESTFRAMEWORK_ENTRY_POINT_END() \
-    return 0;                                \
-    }
-
-#else
-#  define EZ_TESTFRAMEWORK_ENTRY_POINT_END()                        \
-    while (ezTestSetup::RunTests() == ezTestAppRun::Continue)       \
-    {                                                               \
-    }                                                               \
-    const ezInt32 iFailedTests = ezTestSetup::GetFailedTestCount(); \
-    ezTestSetup::DeInitTestFramework();                             \
-    return iFailedTests;                                            \
-    }
-
-#endif
-
-
-#define EZ_TESTFRAMEWORK_ENTRY_POINT(szTestName, szNiceTestName)             \
-  EZ_TESTFRAMEWORK_ENTRY_POINT_BEGIN(szTestName, szNiceTestName)             \
-  /* Execute custom init code here by using the BEGIN/END macros directly */ \
-  EZ_TESTFRAMEWORK_ENTRY_POINT_END()
 
 /// \brief Enum for usage in EZ_TEST_BLOCK to enable or disable the block.
 struct ezTestBlock
