+++ Code/UnitTests/TestFramework/Platform/uwp/uwpTestApplication.cpp
